refactor(contracts): move the lease scope vocabulary into @agent-device/contracts
