+++ packages/contracts/src/__tests__/lease-scope.test.ts
