+++ packages/contracts/src/lease-scope.ts
