+++ src/client/client-normalizers.ts
@@ -21,7 +21,10 @@ import type { SessionRuntimeHints } from '@agent-device/kernel/contracts';
 import { AppError, type DaemonError } from '@agent-device/kernel/errors';
 import { sanitizeErrorCause } from '@agent-device/kernel/redaction';
 import type { SnapshotNode } from '@agent-device/kernel/snapshot';
-import { leaseScopeFromOptions, leaseScopeToRequestMeta } from '../core/lease-scope.ts';
+import {
+  leaseScopeFromOptions,
+  leaseScopeToRequestMeta,
+} from '@agent-device/contracts/lease-scope';
 import type { DaemonRequest } from '../daemon/daemon-request.ts';
 import {
   asRecord,
