+++ src/commands/command-flags.ts
@@ -1,6 +1,9 @@
 import type { CommandFlags } from '@agent-device/contracts/command';
 import { screenshotFlagsFromOptions } from '@agent-device/contracts/capture';
-import { leaseScopeFromOptions, leaseScopeToCommandFlags } from '../core/lease-scope.ts';
+import {
+  leaseScopeFromOptions,
+  leaseScopeToCommandFlags,
+} from '@agent-device/contracts/lease-scope';
 import { stripUndefined } from '@agent-device/kernel/record';
 import { getFlagDefinitions } from './cli-grammar/flag-registry.ts';
 import type { InternalRequestOptions } from '@agent-device/contracts/client';
