+++ src/daemon/__tests__/lease-context.test.ts
@@ -6,7 +6,7 @@ import {
   resolveRequestOrSessionLeaseScope,
   type SessionLease,
 } from '../lease-context.ts';
-import { buildLeaseDiagnosticsContext } from '../../core/lease-scope.ts';
+import { buildLeaseDiagnosticsContext } from '@agent-device/contracts/lease-scope';
 import type { DaemonRequest } from '../daemon-request.ts';
 
 test('buildSessionLeaseFromRequest captures complete request lease scope', () => {
